requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
django-cors-headers==4.0.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-http-client==3.3.7
pytz==2023.3
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing-extensions==4.6.3
